Added incomplete hybrid rendering - No search added - Need to change the `generateUrl` in the client-component
